+++ spec/features/staff/users_can_view_an_organisation_spec.rb
@@ -69,7 +69,7 @@ RSpec.feature "Users can view an organisation" do
   end
 
   context "when the user does not belong to BEIS" do
-    let(:organisation) { create(:organisation) }
+    let(:organisation) { create(:delivery_partner_organisation) }
 
     before do
       authenticate!(user: create(:administrator, organisation: organisation))
@@ -79,6 +79,54 @@ RSpec.feature "Users can view an organisation" do
       visit organisation_path(organisation)
 
       expect(page).to have_content(organisation.name)
+    end
+
+    scenario "can see a list of their programmes, grouped by fund" do
+      newton = create(:fund_activity, :newton)
+      newton_funded_programmes = create_list(:programme_activity, 3, parent: newton, extending_organisation: organisation)
+
+      gcrf = create(:fund_activity, :gcrf)
+      gcrf_funded_programmes = create_list(:programme_activity, 2, parent: gcrf, extending_organisation: organisation)
+
+      visit organisation_path(organisation)
+
+      within_table(id: newton.id) do
+        newton_funded_programmes.each do |programme|
+          within(id: programme.id) do
+            expect(page).to have_content(programme.title)
+            expect(page).to have_content(programme.roda_identifier_compound)
+            expect(page).to have_link(t("default.link.view"))
+          end
+        end
+      end
+
+      within_table(id: gcrf.id) do
+        gcrf_funded_programmes.each do |programme|
+          within(id: programme.id) do
+            expect(page).to have_content(programme.title)
+            expect(page).to have_content(programme.roda_identifier_compound)
+            expect(page).to have_link(t("default.link.view"))
+          end
+        end
+      end
+    end
+
+    scenario "clicking the 'view' link goes to the activity details page" do
+      gcrf = create(:fund_activity, :gcrf)
+      programme = create(:programme_activity, parent: gcrf, extending_organisation: organisation)
+
+      visit organisation_path(organisation)
+
+      within(id: programme.id) do
+        click_link t("default.link.view")
+      end
+
+      expected_path = organisation_activity_details_path(organisation, programme)
+      expect(page.current_path).to eq(expected_path)
+
+      within("h1") do
+        expect(page).to have_content(programme.title)
+      end
     end
 
     scenario "does not see a back link on their organisation home page" do